Bits of Kotlin-related cleanup
- Remove vestiges of Groovy code, especially IOUtil.
- Consolidate Gradle API extensions into GradleApi.kt
- Use `null` less
